+++ fuzz/fuzz_targets/sql_parser.rs
@@ -1,0 +1,8 @@
+#![no_main]
+
+use basalt::sql::parser::parse;
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|input: &str| {
+    let _ = parse(input);
+});